A lot
The biggest thing is that i added the Project settings gui, so everything is prepared for the implementation into the AutoComplete of addons.

I also added a Auto saver that loops every 25 seconds, further the code is saved on a Focus loss of the code area.

I noticed the highlighting was throwing a Stackoverflow error when adding quotes at a specific position in a skript.
Wíll have to Check this later

+++ src/main/kotlin/com/skide/core/management/ProjectManager.kt
@@ -38,6 +38,7 @@ class ProjectManager(val coreManager: CoreManager) {
 
 
     }
+
     fun importProject(name: String, path: String, skriptVersion: String, open: Boolean) {
 
 
@@ -51,14 +52,13 @@ class ProjectManager(val coreManager: CoreManager) {
         if (open) openProject(pointHolder)
 
 
-
     }
 
     fun openProject(holder: PointerHolder) {
 
         val projectConfig = loadProjectConfiguration(holder)
 
-        if(projectConfig.first == ProjectConfigurationLoadResult.SUCCESS && projectConfig.second != null) {
+        if (projectConfig.first == ProjectConfigurationLoadResult.SUCCESS && projectConfig.second != null) {
 
             openProjects.addElement(OpenProject(projectConfig.second!!, coreManager))
         }
@@ -74,7 +74,7 @@ class ProjectManager(val coreManager: CoreManager) {
 
             val obj = JSONObject(readResult.second)
 
-            val project = Project(obj.getLong("project_id"), obj.getString("name"), File(holder.path), obj.getString("skript_version"), obj.getJSONArray("files"), obj.getLong("primary_server_id"))
+            val project = Project(obj.getLong("project_id"), obj.getString("name"), File(holder.path), obj.getString("skript_version"), obj.getJSONArray("files"), obj.getJSONArray("addons"), obj.getLong("primary_server_id"))
             return Pair(ProjectConfigurationLoadResult.SUCCESS, project)
 
         }
@@ -92,19 +92,22 @@ class ProjectManager(val coreManager: CoreManager) {
         obj.put("skript_version", skriptVersion)
         obj.put("project_id", id)
         obj.put("files", JSONArray())
+        obj.put("addons", JSONArray())
         obj.put("primary_server_id", -1)
 
         val configFile = File(projectFolder, ".project.skide")
         writeFile(obj.toString().toByteArray(), configFile, false, true)
         return true
     }
+
     private fun createProjectFilesFromImport(name: String, path: String, skriptVersion: String, id: Long): Boolean {
 
-        println(path)
         val projectFolder = File(path)
 
+        if (!projectFolder.exists()) return false
+
         val possOldConfigFile = File(projectFolder, ".project.skide")
-        if(possOldConfigFile.exists()) {
+        if (possOldConfigFile.exists()) {
             //TODO inform user
             possOldConfigFile.delete()
         }
@@ -112,6 +115,7 @@ class ProjectManager(val coreManager: CoreManager) {
         obj.put("name", name)
         obj.put("skript_version", skriptVersion)
         obj.put("project_id", id)
+        obj.put("addons", JSONArray())
         obj.put("primary_server_id", -1)
         val filesArr = JSONArray()
         projectFolder.listFiles().forEach {
@@ -129,18 +133,24 @@ class ProjectFileManager(val project: Project) {
 
     val configFile = File(project.folder, ".project.skide")
     val projectFiles = HashMap<String, File>()
+    val addons = HashMap<String, String>()
+
 
     init {
         project.files.forEach {
             val f = File(it as String)
-            if(f.exists()) {
-                projectFiles.put(f.name, f)
+            if (f.exists()) {
+                projectFiles[f.name] = f
             }
+        }
+        project.addons.forEach {
+            it as JSONObject
+            addons[it.getString("name")] = it.getString("version")
         }
         rewriteConfig()
     }
 
-    private fun rewriteConfig() {
+    fun rewriteConfig() {
 
         val obj = JSONObject()
         obj.put("name", project.name)
@@ -148,36 +158,61 @@ class ProjectFileManager(val project: Project) {
         obj.put("project_id", project.id)
         obj.put("primary_server_id", project.primaryServer)
         val filesArr = JSONArray()
-        for(file in projectFiles.values) {
+        for (file in projectFiles.values) {
             filesArr.put(file.absolutePath)
         }
+        val addonsArray = JSONArray()
+        for (addon in addons) {
+            val addonObj = JSONObject()
+            addonObj.put("name", addon.key)
+            addonObj.put("version", addon.value)
+            addonsArray.put(addonObj)
+        }
+        obj.put("addons", addonsArray)
         obj.put("files", filesArr)
 
         writeFile(obj.toString().toByteArray(), configFile)
     }
 
-    fun addFile(name:String): Boolean {
-        val rName = if(name.contains(".")) name else name + ".sk"
-        if(projectFiles.containsKey(rName)) return false
+    fun addAddon(name: String, version: String) {
+        addons[name] = version
+        rewriteConfig()
+    }
+
+    fun changeAddonVersion(name: String, version: String) {
+        addons[name] = version
+        rewriteConfig()
+    }
+
+    fun removeAddon(name: String) {
+        addons.remove(name)
+        rewriteConfig()
+    }
+
+    fun addFile(name: String): Boolean {
+        val rName = if (name.contains(".")) name else "$name.sk"
+        if (projectFiles.containsKey(rName)) return false
         val file = File(project.folder, rName)
         file.createNewFile()
         projectFiles.put(rName, file)
         rewriteConfig()
         return true
     }
-    fun deleteFile(rName:String): Boolean {
-        if(!projectFiles.containsKey(rName)) return false
+
+    fun deleteFile(rName: String): Boolean {
+        if (!projectFiles.containsKey(rName)) return false
         val file = projectFiles[rName]
-        if(file?.exists()!!) file.delete() else return false
+        if (file?.exists()!!) file.delete() else return false
         projectFiles.remove(rName)
         rewriteConfig()
         return true
     }
-    fun reNameFile(rName:String, newNameRaw:String): Boolean {
-        val newName = if(newNameRaw.contains(".")) newNameRaw else newNameRaw + ".sk"
-        if(!projectFiles.containsKey(rName)) return false
+
+    fun reNameFile(rName: String, newNameRaw: String): Boolean {
+        val newName = if (newNameRaw.contains(".")) newNameRaw else newNameRaw + ".sk"
+        if (!projectFiles.containsKey(rName)) return false
         val file = projectFiles[rName]
-        if(!file?.exists()!!) return false
+        if (!file?.exists()!!) return false
         projectFiles.remove(rName)
         Files.move(file.toPath(), file.toPath().resolveSibling(newName))
         val nFile = File(project.folder, newName)
